add tasks

+++ app/models/spree/currency_rates/configuration.rb
@@ -3,7 +3,7 @@ module Spree
     class Configuration < Spree::Preferences::Configuration
       preference :source, :string
       preference :api_key, :string
-      preference :markup_in_percent, :integer, default: 0
+      preference :markup_in_percent, :decimal, default: 0
 
       def bank
         @bank ||= Spree::CurrencyRates::Bank.new
